fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
authlib==1.3.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator>=2.0.0
langchain==0.1.0
langchain-openai==0.0.5
openai==1.12.0
pillow==10.1.0
moviepy==1.0.3
numpy==1.24.3